fix data type to float32 in img_matrix
Former-commit-id: 7d5293f20a2c2287fc5a5eb2a4a315e27226521c
Former-commit-id: 98f8d46178b0a9b585433d3638bb8b8995d61107
Former-commit-id: 969aab6c26be3423457e0e79c15445ed8e364c2c [formerly 1b637cd8cb18406d23e3906cdc16bc38f54340f8]
Former-commit-id: 96381afb011b3cad3793da7e3077ce5d07d66643

+++ 2_Implementation/py/transient_visualizer/transient_visualizer.py
@@ -244,7 +244,7 @@ def img_matrix(channels, output, out_type, alpha):
     images = []  # Empty list that will contain all the images
     # Fuse the channels together to obtain a proper [A, R, G, B] image
     for i in tqdm(range(np.shape(frame_A)[2])):
-        img = numpy.empty([np.shape(frame_A)[0], np.shape(frame_A)[1], len(channels)])  # Create an empty numpy array of the correct shape
+        img = numpy.empty([np.shape(frame_A)[0], np.shape(frame_A)[1], len(channels)], dtype=np.float32)  # Create an empty numpy array of the correct shape
 
         img[:, :, 0] = frame_R[:, :, i]
         img[:, :, 1] = frame_G[:, :, i]
